+++ src/pydvl/influence/torch/functional.py
@@ -92,7 +92,7 @@ def batch_hvp_gen(
 
     for inputs, targets in iter(data_loader):
         batch_loss = batch_loss_function(model, loss, inputs, targets)
-        model_params = dict(model.named_parameters())
+        model_params = {k: p for k, p in model.named_parameters() if p.requires_grad}
 
         def batch_hvp(vec: TorchTensorContainerType) -> torch.Tensor:
             aligned_params = align_structure(model_params, vec)
@@ -175,10 +175,8 @@ def batch_loss_function(
         A function that computes the loss of the model on the batch for given model parameters.
     """
 
-    def batch_loss(params: TorchTensorContainerType) -> torch.Tensor:
-        outputs = functional_call(
-            model, params, (to_model_device(x, model),), strict=True
-        )
+    def batch_loss(params: Dict[str, torch.Tensor]) -> torch.Tensor:
+        outputs = functional_call(model, params, (to_model_device(x, model),))
         return loss(outputs, y)
 
     return batch_loss
@@ -219,7 +217,9 @@ def get_hvp_function(
     """
 
     params = {
-        k: p if track_gradients else p.detach() for k, p in model.named_parameters()
+        k: p if track_gradients else p.detach()
+        for k, p in model.named_parameters()
+        if p.requires_grad
     }
 
     def hvp_function(vec: torch.Tensor) -> torch.Tensor: