+++ src/pydvl/value/sampler.py
@@ -27,7 +27,7 @@ The samplers are used in the [semivalues][pydvl.value.semivalues] module to
 compute any semi-value, in particular Shapley and Beta values, and Banzhaf
 indices.
 
-# Slicing of samplers
+## Slicing of samplers
 
 The samplers can be sliced for parallel computation. For those which are
 embarrassingly parallel, this is done by slicing the set of "outer" indices and
@@ -36,6 +36,15 @@ samplers, such as [DeterministicUniformSampler][pydvl.value.sampler.Deterministi
 and [UniformSampler][pydvl.value.sampler.UniformSampler]. In contrast, slicing a
 [PermutationSampler][pydvl.value.sampler.PermutationSampler] creates a new
 sampler which iterates over the same indices.
+
+
+## References
+
+[^1]: <a name="mitchell_sampling_2022"></a>Mitchell, Rory, Joshua Cooper, Eibe
+      Frank, and Geoffrey Holmes. [Sampling Permutations for Shapley Value
+      Estimation](http://jmlr.org/papers/v23/21-0439.html). Journal of Machine
+      Learning Research 23, no. 43 (2022): 1–46.
+
 """
 
 from __future__ import annotations
@@ -191,7 +200,7 @@ class PowersetSampler(abc.ABC, Iterable[SampleT], Generic[IndexT]):
         return len(self._outer_indices)
 
     def __str__(self):
-        return f"{self.__class__.__name__}"
+        return self.__class__.__name__
 
     def __repr__(self):
         return f"{self.__class__.__name__}({self._indices}, {self._outer_indices})"
@@ -315,18 +324,19 @@ class AntitheticSampler(StochasticSamplerMixin, PowersetSampler[IndexT]):
     """An iterator to perform uniform random sampling of subsets, and their
     complements.
 
-    Works as :class:`~pydvl.value.sampler.UniformSampler`, but for every tuple
-    $(i,S)$, it subsequently returns $(i,S^c)$, where $S^c$ is the complement of
-    the set $S$, including the index $i$ itself.
+    Works as [UniformSampler][pydvl.value.sampler.UniformSampler], but for every
+    tuple $(i,S)$, it subsequently returns $(i,S^c)$, where $S^c$ is the
+    complement of the set $S$ in the set of indices, excluding $i$.
     """
 
     def __iter__(self) -> Iterator[SampleT]:
         while True:
             for idx in self.iterindices():
-                subset = random_subset(self.complement([idx]), seed=self._rng)
+                _complement = self.complement([idx])
+                subset = random_subset(_complement, seed=self._rng)
                 yield idx, subset
                 self._n_samples += 1
-                yield idx, self.complement(np.concatenate((subset, np.array([idx]))))
+                yield idx, np.setxor1d(_complement, subset)
                 self._n_samples += 1
             if self._n_samples == 0:  # Empty index set
                 break
@@ -370,6 +380,29 @@ class PermutationSampler(StochasticSamplerMixin, PowersetSampler[IndexT]):
     @classmethod
     def weight(cls, n: int, subset_len: int) -> float:
         return n * math.comb(n - 1, subset_len) if n > 0 else 1.0
+
+
+class AntitheticPermutationSampler(PermutationSampler[IndexT]):
+    """Samples permutations like
+    [PermutationSampler][pydvl.value.sampler.PermutationSampler], but after
+    each permutation, it returns the same permutation in reverse order.
+
+    This sampler was suggested in (Mitchell et al. 2022)<sup><a
+    href="#mitchell_sampling_2022">1</a></sup>
+
+    !!! tip "New in version 0.7.1"
+    """
+
+    def __iter__(self) -> Iterator[SampleT]:
+        while True:
+            permutation = self._rng.permutation(self._indices)
+            for perm in permutation, permutation[::-1]:
+                for i, idx in enumerate(perm):
+                    yield idx, perm[:i]
+                    self._n_samples += 1
+
+            if self._n_samples == 0:  # Empty index set
+                break
 
 
 class DeterministicPermutationSampler(PermutationSampler[IndexT]):