+++ tests/conftest.py
@@ -25,10 +25,10 @@ if TYPE_CHECKING:
 
 def pytest_addoption(parser):
     parser.addoption(
-        "--do-not-start-memcache",
+        "--memcached-service",
         action="store_true",
-        help="When this flag is used, memcache won't be started by a fixture"
-        " and is instead expected to be already running",
+        default="localhost:11211",
+        help="Address of memcached server to use for tests.",
     )
     group = parser.getgroup("tolerate")
     group.addoption(
@@ -101,64 +101,17 @@ def is_memcache_responsive(hostname, port):
 
 
 @pytest.fixture(scope="session")
-def do_not_start_memcache(request):
-    return request.config.getoption("--do-not-start-memcache")
-
-
-@pytest.fixture(scope="session")
-def docker_services(
-    docker_compose_command,
-    docker_compose_file,
-    docker_compose_project_name,
-    docker_setup,
-    docker_cleanup,
-    do_not_start_memcache,
-):
-    """Start all services from a docker compose file (`docker-compose up`).
-    After test are finished, shutdown all services (`docker-compose down`)."""
-    from pytest_docker.plugin import get_docker_services
-
-    if do_not_start_memcache:
-        yield
-    else:
-        with get_docker_services(
-            docker_compose_command,
-            docker_compose_file,
-            docker_compose_project_name,
-            docker_setup,
-            docker_cleanup,
-        ) as docker_service:
-            yield docker_service
-
-
-@pytest.fixture(scope="session")
-def memcached_service(docker_ip, docker_services, do_not_start_memcache):
-    """Ensure that memcached service is up and responsive.
-
-    If `do_not_start_memcache` is True then we just return the default values
-    'localhost', 11211
-    """
-    if do_not_start_memcache:
-        return "localhost", 11211
-    else:
-        # `port_for` takes a container port and returns the corresponding host port
-        port = docker_services.port_for("memcached", 11211)
-        hostname, port = docker_ip, port
-        docker_services.wait_until_responsive(
-            timeout=30.0,
-            pause=0.5,
-            check=lambda: is_memcache_responsive(hostname, port),
-        )
-        return hostname, port
+def memcached_service(request) -> tuple[str, int]:
+    opt = request.config.getoption("--memcached-service", default="localhost:11211")
+    host, port = opt.split(":")
+    return host, int(port)
 
 
 @pytest.fixture(scope="function")
 def memcache_client_config(memcached_service) -> MemcachedClientConfig:
-    client_config = MemcachedClientConfig(
+    return MemcachedClientConfig(
         server=memcached_service, connect_timeout=1.0, timeout=1, no_delay=True
     )
-    Client(**asdict(client_config)).flush_all()
-    return client_config
 
 
 @pytest.fixture(scope="function")
@@ -170,11 +123,9 @@ def memcached_client(memcache_client_config) -> Tuple[Client, MemcachedClientCon
         c.flush_all()
         return c, memcache_client_config
     except Exception as e:
-        print(
-            f"Could not connect to memcached server "
-            f'{memcache_client_config["server"]}: {e}'
-        )
-        raise e
+        raise ConnectionError(
+            f"Could not connect to memcached at {memcache_client_config.server}"
+        ) from e
 
 
 @pytest.fixture(scope="function")
@@ -261,6 +212,14 @@ def pytest_configure(config: "Config"):
 def tolerate(request: pytest.FixtureRequest):
     fixture = TolerateErrorFixture(request.node)
     return fixture
+
+    worker_id = os.environ.get("PYTEST_XDIST_WORKER")
+    if worker_id is not None:
+        logging.basicConfig(
+            format="%(asctime)s %(levelname)s %(message)s",
+            filename=f"tests_{worker_id}.log",
+            level=logging.DEBUG,
+        )
 
 
 def pytest_runtest_setup(item: pytest.Item):