+++ tests/utils/test_numeric.py
@@ -5,6 +5,7 @@ from pydvl.utils.numeric import (
     powerset,
     random_matrix_with_condition_number,
     random_powerset,
+    random_powerset_label_min,
     random_subset_of_size,
     running_moments,
 )
@@ -248,3 +249,27 @@ def test_running_moments():
         true_variances = [np.var(vv) for vv in values]
         assert np.allclose(means, true_means)
         assert np.allclose(variances, true_variances)
+
+
+@pytest.mark.parametrize(
+    "min_elements_per_label,num_elements_per_label,num_labels,check_num_samples",
+    [(0, 10, 3, 1000), (1, 10, 3, 1000), (2, 10, 3, 1000)],
+)
+def test_random_powerset_label_min(
+    min_elements_per_label: int,
+    num_elements_per_label: int,
+    num_labels: int,
+    check_num_samples: int,
+):
+    s = np.arange(num_labels * num_elements_per_label)
+    labels = np.arange(num_labels).repeat(num_elements_per_label)
+
+    for idx, subset in enumerate(
+        random_powerset_label_min(s, labels, min_elements_per_label)
+    ):
+        assert np.all(np.isin(subset, s))
+        for group in np.unique(labels):
+            assert np.sum(group == labels[subset]) >= min_elements_per_label
+
+        if idx == check_num_samples:
+            break